- Changes for DM37LINUX-446:  - Modify min/max computaiton in CONFIG_SPLASH_SCREEN_ALIGN section to be    signed; necessary if logi is taller/wider than image.  - Clip log down to 332x57 pixels.  - Modify DSS setup to use actual Hz in clock calculations instead of KHz    for more precise calculation.  Allow fclk_div to go to fck_min_div; it    was one short which caused some calculations to not get as close of a    pixel clock as possible.  - result_fck_div is best_dss.fck_div - using current value may not be correct.  - Rename find_panel to be logic_find_panel and make non-static.
git-svn-id: svn+ssh://svn/srv/svn/eps_svn/software/products/linux/u-boot/u-boot-2011.06@26242 d7e73fac-2431-0410-abd5-fd325483733a

+++ common/lcd.c
@@ -1003,14 +1003,14 @@ int lcd_display_bitmap(ulong bmp_image, int x, int y)
 
 #ifdef CONFIG_SPLASH_SCREEN_ALIGN
 	if (x == BMP_ALIGN_CENTER)
-		x = max(0, (pwidth - width) / 2);
+		x = max(0, ((int)pwidth - (int)width) / 2);
 	else if (x < 0)
-		x = max(0, pwidth - width + x + 1);
+		x = max(0, (int)pwidth - (int)width + x + 1);
 
 	if (y == BMP_ALIGN_CENTER)
-		y = max(0, (panel_info.vl_row - height) / 2);
+		y = max(0, ((int)panel_info.vl_row - (int)height) / 2);
 	else if (y < 0)
-		y = max(0, panel_info.vl_row - height + y + 1);
+		y = max(0, (int)panel_info.vl_row - (int)height + y + 1);
 
 #endif /* CONFIG_SPLASH_SCREEN_ALIGN */
